Tests, serializers and validators improvements
- Removed useless function from BillValidator
- Bill fields optimized to a nice view
- Created a folder and new files to tests
- Removed old tests file
- Added new log folder
- Get environment debug variable
- Setted status code 422 to GetBillByMonth exception
- Change .coveragerc to not cov tests files
