+++ .travis.yml
@@ -14,8 +14,8 @@ before_script:
 - psql -c 'create database test_db;' -U postgres
 
 script:
-- python manage.py test
-- coverage run manage.py test
+- python manage.py test callcenter/tests/
+- coverage run manage.py test callcenter/tests/
 
 after_success:
 - codecov --gcov-glob callcenter/schemas.py
